# docs/api/openapi-mvp.yaml
openapi: 3.0.3
info:
  title: Moimday API
  version: 1.3.0
  description: MVP REST API ??SSOT with api-contract-mvp.md (multi-group, 30 members)
servers:
  - url: http://localhost:3000/v1
    description: Local dev
  - url: https://api.Moimday.example/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /auth/oauth/{provider}:
    post:
      tags: [auth]
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [kakao, google, apple]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                idToken: { type: string, description: google | apple }
                accessToken: { type: string, description: kakao }
      responses:
        '200':
          description: Tokens + user
        '401': { description: UNAUTHORIZED }
        '503': { description: SERVICE_UNAVAILABLE }
  /auth/token/refresh:
    post:
      tags: [auth]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        '200': { description: New access token }
...